requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.4
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
exceptiongroup==1.1.0
iniconfig==2.0.0
Markdown==3.4.1
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
